lots of functionality & final(ish) color-> month mapping

+++ 1.mapping-time/students/michael/sketch.js
@@ -1,20 +1,30 @@
 /* FEATURES
-Stroke object
-Transforms / Rotate
-Perlin noise on pen strokes
-TODO: timer events 
-TODO: Slow reaveal strokes
-TODO: Slow color strokes
-TODO: Stroke death: fade, wind?
+Done: Stroke object
+Done: Transforms / Rotate
+Done: Perlin noise on pen strokes
+Done: timer events 
+Done: Slow reveal strokes
+Done: Slow color strokes
+Done: array add / remove 
+zero point offset - taken care of w sDisp
+Done: mouseSplosion
+Done: stroke: off by 1
 
-TODO: Blot object
+TODO: leaf sizing
+TODO: ordered drawing outside to inside
 TODO: Meandering pen movement
-
+TODO: Blot object
 TODO: csv reader - decided what == distraction
 TODO: resize update propagation
-TODO: array add / remove 
-TODO: get tint() working
-zero point offset - taken care of w sDisp
+
+Extra
+TODO: sound
+TODO: make sec dots at center instead
+TODO: every x sec is bold?
+TODO: make stroke more ornate
+TODO: Stroke death: fade, wind?
+TODO: get tint() working / nope. broken but good.
+
 */
 
 /* Visualizations
@@ -46,13 +56,16 @@ function Stroke (color, type, val, month) {
     this.type = type;
     this.locX = midX;
     this.locY = midY;
+    this.endX = 0;
+    this.endY = 0;
     this.rotation = 0;
     this.g = createGraphics(gX, gY);
     this.sFrame = frameCount;	// start fc
     this.eFrame = -1;		// end fc
     
     this.genStroke = function (cPercent) {
-      // add self to array
+      this.g.tint(55, 4);  // not working
+
       var x=0;
       var y=0;
       var c = lerpColor(black, this.color, cPercent);	// start off like black ink
@@ -73,24 +86,29 @@ function Stroke (color, type, val, month) {
       }
     }
     
-    this.live = function () {
-      
-    }
-    
-    this.die = function () {
-      
-    }
-    
     // array filter test
     this.fTest  = function (val) {
       return (val != this);
     }
     
     this.update = function () {
-      if (frameCount-this.sFrame <= fadeTime)  // fade in
+      
+      if (this.alive) {  // grow in-out
+        if (this.endX < gX) this.endX+=gX/4;
+        if (this.endY < gY) this.endY+=gY/4;
+      } else {
+        if (this.endX >= 0) this.endX-=gX/4;
+        if (this.endY >= 0) this.endY-=gY/4;
+        if (this.endX <= 0 || this.endY <= 0) { // sanity check
+          this.endX = 2;
+          this.endY = 2;
+        }
+      }
+    
+      if (frameCount-this.sFrame <= fadeTime) {  // fade in
         this.genStroke((frameCount-this.sFrame) / fadeTime);
-        
-      if (! this.alive && this.eFrame < 0) 
+      }
+      if (! this.alive && this.eFrame < 0) // set death timer
         this.eFrame = frameCount;
         
       if (this.eFrame > 0 && frameCount-this.eFrame <= (fadeTime/1.5)) { // fade out
@@ -112,22 +130,78 @@ function Stroke (color, type, val, month) {
       translate(midX, midY);
       rotate(angleStep * val * sDisp);
       
-      // tint(55, 100, 46, 4);  // not working
       mouseSplode = map(mouseX, 0, width, 1, 6);
-      image(this.g, sDens*this.type*mouseSplode, sDens*this.type*mouseSplode);
+      drawStart = sDens*this.type*mouseSplode;
+      image(this.g, drawStart, drawStart, this.endX, this.endY);
       pop()
     }
 }
 
+Date.prototype.isLeapYear = function() {
+    var year = this.getFullYear();
+    if((year & 3) != 0) return false;
+    return ((year % 100) != 0 || (year % 400) == 0);
+};
+
+// Get Day of Year
+Date.prototype.getDOY = function() {
+    var dayCount = [0, 31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334];
+    var mn = this.getMonth();
+    var dn = this.getDate();
+    var dayOfYear = dayCount[mn] + dn;
+    if(mn > 1 && this.isLeapYear()) dayOfYear++;
+    return dayOfYear;
+};
+
 currColor = [];
 nextColor = [];
 
-function setup() {
-  currColor = [color(0, 0, 255, 20), color(255, 0, 0, 20), color(0, 255, 0, 20), color(0, 255, 255, 20)];
-  nextColor = [color(0, 0, 255, 20), color(255, 0, 0, 20), color(0, 255, 0, 20), color(0, 255, 255, 20)];
-  tMax = [60, 60, 24, 32 - new Date(year(), month(), 32).getDate(), 12];
+// set lerping colors by day of year
+function setSeason () {
+  d = new Date;
+  doy = d.getDOY();
+  currColor = winterColor;
+  nextColor = springColor;
   
-  black = color(0);
+  if (doy > 79){ // BUG: leap year
+    currColor = springColor;
+    nextColor = summerColor;
+  }
+  if (doy > 172){
+    currColor = summerColor;
+    nextColor = fallColor;
+  }
+  if (doy > 265){
+    currColor = fallColor;
+    nextColor = winterColor;
+  }
+  
+  if (doy > 355){
+    currColor = winterColor;
+    nextColor = springColor;
+  }  
+}
+
+function setup() {
+  /* seasonal colors:
+    summer	= bright egg yolk
+    fall	= maple leaf orange/red
+    winter	= aquamarine
+    spring	= lime green
+  */
+  
+  //summerColor = [color('#ffd827'), color('#fe6b05'), color('#ec1b45'), color('#668013')];
+  summerColor = [color('#f8ff00'), color('#fe6b05'), color('#c20065'), color('#067515')];
+  fallColor = [color('#7e0304'), color('#d33f0f'), color('#fb8840'), color('#fcaf67')];
+  winterColor = [color('#b5a5bf'), color('#d8cce0'), color('#757da4'), color('#13161f')];
+  //springColor = [color('#dbec90'), color('#c5db5d'), color('#7b9677'), color('#11b99f')]; 
+  springColor = [color('#f8e69a'), color('#fab5a6'), color('#96cf4c'), color('#009900')]; 
+
+  setSeason();
+  
+  tMax = [59, 59, 23, 32 - new Date(year(), month(), 32).getDate(), 12];
+  
+  black = color(0, 100);
   
   update();
   createCanvas(windowWidth, windowHeight);
@@ -137,10 +211,9 @@ function setup() {
   t = [second(), minute(), hour(), day(), month()];
   tPrev = t; // previous time
   for (var tidx=0; tidx < t.length - 1; tidx++) {	// make all strokes
-    for (var nidx=0; nidx < t[tidx]; nidx++) {
+    for (var nidx=0; nidx < t[tidx] - 1; nidx++) {
       var c = lerpColor(currColor[tidx], nextColor[tidx], .33);
       var o = new Stroke(c, tidx, nidx, 0)
-      //o.setup();
       sArr.push(o);		// add object to array
     }
   }
@@ -155,12 +228,12 @@ function update () {
   
   for (var idx=0; idx < t.length; idx++) {		// compare all except month
     
-    if (t[idx] != tPrev[idx]) {				// if new...
+    if (t[idx] != tPrev[idx] && t[idx] != 0) {		// if new...
       if (idx == 4) {					// if the month turns over get new end day
         tMax = [60,60,24, 32 - new Date(year(), month(), 32).getDate(), 12];
         continue;
       }
-      if (t[idx] == 0) {					// prevcycle finished
+      if (t[idx] == 0) {					// prev cycle finished
         for (var sidx=0; sidx < sArr.length; sidx++) {		// run through
           if (sArr[sidx].type == idx) {				// if correct type
             sArr[sidx].alive = false;				// kill all strokes for type
@@ -198,7 +271,7 @@ function draw() {
   }
   */
   
-  background(55);
+  background(255);
   fill(255,0,0);
   ellipse(
     200+100*cos(map(millis()%5000,0,5000,0,TWO_PI)),
